sample-code/examples/python/pytest/requirements.txt
Appium-Python-Client
pytest